+++ module-bb-graph/src/main/java/ch/hsr/adv/lib/graph/logic/GraphBuilder.java
@@ -3,6 +3,10 @@ package ch.hsr.adv.lib.graph.logic;
 import ch.hsr.adv.commons.core.logic.domain.Module;
 import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.graph.logic.ConstantsGraph;
+import ch.hsr.adv.commons.graph.logic.domain.ADVEdge;
+import ch.hsr.adv.commons.graph.logic.domain.ADVVertex;
+import ch.hsr.adv.commons.graph.logic.domain.GraphElement;
+import ch.hsr.adv.commons.graph.logic.domain.GraphRelation;
 import ch.hsr.adv.lib.core.logic.ADVModule;
 import ch.hsr.adv.lib.core.logic.Builder;
 import ch.hsr.adv.lib.graph.logic.domain.*;
